Added unit tests for new physics_engine methods, refactored object state classes
